medicine_assistant_package/requirements_assistant.txt
# requirements.txt - Add these to your existing requirements

# OpenAI API
openai>=1.12.0

# Environment variables management
python-dotenv>=1.0.0

# For better JSON handling (optional)
ujson>=5.9.0

# For async operations (if needed)
aiohttp>=3.9.0

# Rate limiting (optional, for production)
django-ratelimit>=4.1.0
